database/requirements.txt
astropy==4.0.1.post1
mysql-connector-python==8.0.21
numpy==1.19.1
protobuf==3.12.2
six==1.15.0
